refactor: move Tag component to top level features directory
