+++ apps/web/app/_features/articles/index.ts
@@ -3,4 +3,6 @@ export { getArticlesMeta } from "./getArticlesMeta";
 /** @public */
 export { getArticleMeta } from "./getArticleMeta";
 /** @public */
+export { Tag } from "./Tag";
+/** @public */
 export type { ArticleMeta, TagEnum } from "./type";
