+++ apps/web/app/articles/_features/ArticleListItem.tsx
@@ -1,8 +1,7 @@
 import { UniversalLink } from "../../_components";
 
-import { Tag } from "./Tag";
-
 import type { ArticleMeta } from "@/app/_features";
+import { Tag } from "@/app/_features";
 import { format } from "@/app/_utils";
 
 type Props<T extends string> = ArticleMeta<T>;
